updates to the physics Engine

+++ src/Vector2.cpp
@@ -3,7 +3,7 @@
 #include <math.h>
 
 
-Vector2::Vector2(int x,int y)
+Vector2::Vector2(float x,float y)
 {
     this->x = x;
     this->y = y;
