+++ src/Vector2.h
@@ -10,7 +10,7 @@ class Vector2
 public:
     float x = 0;
     float y = 0;
-    Vector2(int x,int y);
+    Vector2(float x,float y);
     Vector2();
     inline Vector2 operator + (Vector2 v2)
     {
